Fixed regressions

+++ HJM.Chip8/HJM.Chip8.CPU/Instructions/CLS_00E0.cs
@@ -19,7 +19,7 @@ namespace HJM.Chip8.CPU.Instructions
             {
                 stateChange.GraphicsChanges.Add(new AddressChange<byte>
                 {
-                    AddressChanged = (byte)i,
+                    AddressChanged = i,
                     OldValue = state.Graphics[i],
                     NewValue = 0
                 });
